requirements.txt
langchain-google-genai==4.1.2
langchain==1.2.0
mem0ai==1.0.1
python-dotenv==1.2.1